Improve BST printing
The previous function was invalid when it came to printing level
delimiters. There's a simpler way to track the delimiter: when
printing a delimiter and there are stuff in the Q (I'm using a slice
and an index cause I'm lazy) we need to add another delimiter cause
we're at the end of the next level.

Like everything with tree it's easier to explain graphically than with
words but here we are... in a commit message.