Implement an autocomplete
It's more fun to work with data structures in an actual application.
This one is a very simple autocomplete API.
It can be improve by adding ranks to word nodes. Then the returned
strings could be sorted by rank. And probably limited by some number
like top 10. Then it will be more like the real thing. But the
interesting part was to traverse the trie to collect the words.